requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.12
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
scikit-learn>=1.3.0
pyyaml>=6.0
python-dotenv>=1.0.0
openpyxl>=3.1.0
requests>=2.31.0
jupyter>=1.0.0
streamlit>=1.28.0
fastapi>=0.100.0
uvicorn>=0.23.0